Определил основные классы
