+++ ToFrends/IUser.cs
@@ -1,0 +1,22 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace ToFrends
+{
+
+    interface IUser
+    {
+        string Name { get; set; }
+        string Adress { get; set; }
+        string Login { get; set; }
+        string Pass { get; set; }
+
+    }
+
+
+
+
+}
